// lib/enums.ts
export enum TimeFrame {
  ASAP = "asap",
  RANDOM = "random",
  SPECIFIC = "specific",
}

export enum ItemCondition {
  NEW = "new",
  USED = "used",
  DAMAGED = "damaged",
}

export enum BookingStatus {
  PENDING = "pending",
  ACCEPTED = "accepted",
  DECLINED = "declined",
}

export enum NotificationType {
  CHAT = "chat",
  BOOKING = "booking",
  REQUEST = "request",
}